% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gdal_sd.R
\name{gdal_sd}
\alias{gdal_sd}
\title{Cell-wise raster standard deviation}
\usage{
gdal_sd(infile, outfile, return_raster = TRUE, quiet = TRUE)
}
\arguments{
\item{infile}{A file path to a single, multiband raster file, or a vector of
paths to multiple raster files. If a vector of paths is provided, then if
any of the rasters have multiple bands, the first band will be used. A
maximum of 26 bands or files can be provided to \code{infile}.}

\item{outfile}{Optional. A path to the desired output raster file, which will
be created if it doesn't already exist (though the containing directory
must exist). If not provided and \code{return_raster} is \code{TRUE}, a
temporary output file will be created.}

\item{return_raster}{Logical. Should the output raster be read back in to R
as a \code{Raster} object?}

\item{quiet}{Logical. Should GDAL output messages be suppressed?}
}
\value{
Unless \code{return_raster} is \code{TRUE}, \code{gdal_sd} creates
  the file specified at \code{outfile} but returns \code{NULL}, invisibly, to
  R. If \code{return_raster} is \code{TRUE}, the standard deviation
  \code{Raster} object is returned to R.
}
\description{
Calculate a raster layer whose cells' values are the standard deviations
across the bands of a single raster, or across multiple rasters.
}
\details{
The quantity calculated is the corrected sample standard deviation,
  calculated as \eqn{sqrt(sum((x - mean(x))^2)/(length(x)-1))}.
}
\examples{
library(raster)
s <- stack(replicate(5, raster(matrix(rnorm(100), ncol=10))))

# Standard deviation across multiple files
writeRaster(s, ff <- paste0(tempfile(), 1:5, '.tif'), bylayer=TRUE)
gdal_sd(ff)

# Standard deviation across multiple bands
writeRaster(s, f <- tempfile(fileext='.tif'))
gdal_sd(f)
}

