% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ibra_region.R
\name{ibra_region}
\alias{ibra_region}
\title{Identify IBRA 7 regions}
\usage{
ibra_region(pts, tabulate = TRUE)
}
\arguments{
\item{pts}{A SpatialPolygons* object.}
}
\value{
A vector of IBRA 7 region names corresponding to the IBRA 7 regions
  that a set of points falls within. If any points do not overlie an IBRA 7
  region, these will be ignored with a warning.
}
\description{
Return the names of the set of IBRA regions within which a set of points
falls.
}
\examples{
xy <- SpatialPoints(data.frame(x=runif(1000, 140, 145), y=runif(1000, -38, -35)))
proj4string(xy) <- '+init=epsg:4283'
reg <- ibra_region(xy)
reg
}
\keyword{ibra}
\keyword{spatial,}

