% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interpolate_linear.R
\name{interpolate_linear}
\alias{interpolate_linear}
\title{Linear interpolation between pairs of rasters}
\usage{
interpolate_linear(s, z_in, z_out, outdir, prefix, progress = TRUE,
  write_change = TRUE, return_stack = FALSE, ...)
}
\arguments{
\item{s}{A \code{RasterStack} or \code{RasterBrick} containing the rasters
between which interpolation will be performed.}

\item{z_in}{A numeric vector indicating the positions (e.g., times)
associated with layers in \code{s} (in the same order as the layers of
\code{s} - see \code{names(s)}).}

\item{z_out}{a numeric vector that indicates the times to be interpolated to
(if \code{z_out} extends beyond the latest time slice in \code{z_in}, it
will be extrapolated using the rate of change from the period between the
last and second-to-last time in \code{z_in}).}

\item{outdir}{The directory to which files will be written (recursively
created if not already in existence) (character string).}

\item{prefix}{The output files will have pattern \code{prefix_x.tif}, where
\code{x} is the position (timestep) (potentially multiple digits), and
\code{prefix} is a string that you specify here (character string).}

\item{progress}{Logical. Show progress bar?}

\item{write_change}{Logical. Should the change grids (which define the change
in cells' values per timestep between each pair of time slices) be written
to \code{outdir}.}

\item{return_stack}{Logical. Should the interpolated grids (at timesteps
\code{z_out}) be returned as a \code{RasterStack}?}

\item{...}{Additional arguments passed to writeRaster.}
}
\value{
Unless \code{return_stack} is \code{TRUE}, \code{interpolate}
  creates files in \code{outdir}, but returns \code{NULL}, invisibly, to R.
  If \code{return_stack} is \code{TRUE}, a \code{RasterStack} of
  interpolated \code{Raster} objects is returned to R.
}
\description{
Linearly interpolate cell values between pairs of raster grids, at specified
locations (e.g. "times"). Extrapolation to times beyond the last raster in
the series is also possible.
}
\details{
Extrapolation below the earliest grid in \code{s} is currently not
  implemented. Interpolation is linear between pairs of time slices existing
  in \code{z_in}, and extrapolation beyond the last slice in \code{s}, if
  requested, follows the trend observed between the last and second-to-last
  slice in \code{z_in}.
}
\examples{
library(raster)
r1 <- raster(matrix(rnorm(100), ncol=10))
r2 <- raster(matrix(rnorm(100, mean=11), ncol=10))
s <- stack(r1, r2)
s_interp <- interpolate_linear(s, c(2001, 2011), 2001:2020, tempdir(),
                               'example', return_stack=TRUE)
plot(s_interp, zlim=range(values(s_interp)))
}

