% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nodata.R
\name{nodata}
\alias{nodata}
\title{Query raster files for their NoData values}
\usage{
nodata(srcfile, simplify = TRUE)
}
\arguments{
\item{srcfile}{A character vector of file paths to the raster files to be
queried.}

\item{simplify}{Logical. If \code{TRUE}, then when all files referred to in
\code{srcfile} have the same number of bands, the returned value will be
simplified to an array (if \code{srcfile} refers to more than one file), or
an atomic vector (if \code{srcfile} refers to a single file).}
}
\value{
A list with as many elements as the number of files specified in
  \code{srcfile}. Each list element contains a single vector whose values are
  the NoData values for each band of the corresponding raster file. If GDAL
  is unable to identify the NoData value, \code{NA} will be returned.
}
\description{
Return the NoData value associated with one or more raster files.
}
\examples{
library(raster)
r <- raster(matrix(runif(100), ncol=10))
NAvalue(r) <- -9999
writeRaster(r, f <- tempfile(fileext='.tif'))
nodata(f)
nodata(c(f, f, f))
}
\keyword{query}
\keyword{raster,}
\keyword{spatial,}

