% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proj_by_name.R
\name{proj_by_name}
\alias{proj_by_name}
\title{Get proj4 strings and EPSG codes by CRS keywords.}
\usage{
proj_by_name(x, type = "proj4", crs = FALSE)
}
\arguments{
\item{x}{Character. Key words to search for.}

\item{type}{Either \code{'epsg'} or \code{'proj4'}.}

\item{crs}{Logical. Should a CRS object be returned?}
}
\value{
If \code{crs} is \code{TRUE}, a \code{CRS} object. Otherwise, if
  \code{type} is \code{'epsg'}, a vector of EPSG codes, or if \code{type} is
  \code{'proj4'}, a vector of proj4 strings.
}
\description{
Get proj4 strings and EPSG codes by CRS keywords.
}
\examples{
proj_by_name('gda94', type='proj4')
proj_by_name('gda94 albers', crs=TRUE)
}
\keyword{gis}
\keyword{proj4,}
\keyword{projection,}
\keyword{spatial,}

