% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_mxe.R
\name{read_mxe}
\alias{read_mxe}
\title{Read raster data stored in Maxent's mxe format}
\usage{
read_mxe(file, return_raster = TRUE)
}
\arguments{
\item{file}{The path to the mxe file to be read.}

\item{return_raster}{Logical. If \code{FALSE}, then the cell values data will
returned as a vector; if \code{TRUE}, a \code{raster} object will be
returned.}
}
\value{
If \code{return_raster} is \code{TRUE}, a \code{raster} object. If
\code{return_raster} is \code{FALSE}, a list is returned, with the following
elements:
  \itemize{
  \item{\code{xll}}{: The x coordinate of the lower left corner of the extent.}
  \item{\code{yll}}{: The y coordinate of the lower left corner of the extent.}
  \item{\code{cellsize}}{: The grid resolution. Note that horizontal and
  vertical resolution are assumed equal.}
  \item{\code{nrow}}{: The number of rows of data.}
  \item{\code{ncol}}{: The number of columns of data.}
  \item{\code{nodata}}{: The nodata value.}
  \item{\code{datatype}}{: A character string indicating the data type.}
  \item{\code{data}}{: A vector whose elements are the raster cell values.}
}
}
\description{
Read binary mxe files created by the Maxent habitat suitability modelling
software package.
}
\author{
John B. Baumgartner, \email{johnbaums@gmail.com}

Peter D. Wilson
}
\references{
Based on \href{http://rpubs.com/puddleduck/91946}{"Reading mxe
  files with R - revisited"} by Peter D. Wilson.
}
\seealso{
\code{\link{project_maxent}}
}
\keyword{maxent,}
\keyword{read}

