% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/things-package.R
\docType{package}
\name{things}
\alias{things}
\alias{things-package}
\title{things: Miscellaneous R functions - mainly spatial and SDM-related}
\description{
things: Miscellaneous R functions - mainly spatial and SDM-related
}

