% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjacent_ibra.R
\name{adjacent_ibra}
\alias{adjacent_ibra}
\title{Subset IBRA 7 by points}
\usage{
adjacent_ibra(pts)
}
\arguments{
\item{pts}{A SpatialPolygons* object.}
}
\value{
A SpatialPolygonsDataFrame representing the subset of polygons of
  ibra7_albers within which one or more points (pts) fall, and those polygons
  adjacent to these.
}
\description{
Subset IBRA 7 to those polygons that contain, or are adjacent to polygons
that contain, one or more given points.
}
\examples{
xy <- SpatialPoints(data.frame(x=runif(1000, 140, 145), y=runif(1000, -38, -35)))
proj4string(xy) <- '+init=epsg:4283'
adj <- adjacent_ibra(xy)
plot(adj, col='gray80', border='transparent')
}
\keyword{ibra}
\keyword{spatial,}

