% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdallocationinfo.R
\name{gdallocationinfo}
\alias{gdallocationinfo}
\title{Query raster files at points}
\usage{
gdallocationinfo(srcfile, pts, simplify = TRUE, sp = FALSE)
}
\arguments{
\item{srcfile}{A character vector of file paths to the raster files to be 
queried.}

\item{pts}{A \code{matrix}, \code{data.frame}, or \code{SpatialPoints*} 
object giving the x and y coordinates of a point (or points), given in the 
coordinate reference system of the rasters indicated at \code{srcfile}.}

\item{simplify}{Logical. If \code{TRUE}, then if all files referred to in 
\code{srcfile} are single band rasters, the sampled values will be
simplified to a matrix.}

\item{sp}{Logical. If \code{TRUE}, a \code{SpatialPointsDataFrame} will be
returned, otherwise a \code{data.frame} will be returned. If \code{pts} is
a \code{SpatialPoints*} object and has a non-\code{NA} CRS, then if 
\code{sp} is \code{TRUE}, the returned object will have that same CRS 
assigned.}
}
\value{
If only one file is queried, then a matrix is returned, with columns
  corresponding to bands of the file. If \code{simplify} is \code{TRUE} and
  all files are single band, the returned value will be a \code{matrix} with
  one column for each file, and rows giving the values of cells corresponding
  to the coordinates given by \code{x} and \code{y}. If \code{simplify} is
  \code{FALSE}, or if one or more files is multiband, the returned value is a
  list of matrices, with the columns of each matrix corresponding to the
  bands of the file. Further, if \code{sp} is \code{TRUE}, data will be
  returned as \code{SpatialPointsDataFrames}.
}
\description{
Return cell values for one or more raster files, at a given point or set of 
points. Requires that GDAL is installed and avilable to \code{system} (i.e.
listed in the PATH environmental variable).
}
\examples{
library(raster)
r <- raster(matrix(runif(1.6e7), ncol=4000))
r[sample(ncell(r), 10000000)] <- NA
writeRaster(r, f <- tempfile(fileext='.tif'))
file.info(f)
xy <- data.frame(x=runif(10000), y=runif(10000))

# Querying a single file at xy
vals <- gdallocationinfo(f, pts=xy)

# Querying multiple files (here identical) at xy:
vals2 <- gdallocationinfo(c(f, f, f), pts=xy)

# Returning as a SpatialPointsDataFrame
vals2 <- gdallocationinfo(c(f, f, f), pts=xy, sp=TRUE)
}
\keyword{extract,}
\keyword{points,}
\keyword{query,}
\keyword{raster,}
\keyword{sample}
\keyword{spatial,}

